git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/boblight@28436 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/cmyth/librefmem/alloc.c
@@ -117,7 +117,7 @@ ref_remove(refcounter_t *ref)
 	int bin;
 	refcounter_t *r, *p;
 
-	bin = ((unsigned long)ref >> 2) % REF_ALLOC_BINS;
+	bin = ((uintptr_t)ref >> 2) % REF_ALLOC_BINS;
 
 	r = ref_list[bin];
 	p = NULL;
@@ -141,7 +141,7 @@ ref_add(refcounter_t *ref)
 {
 	int bin;
 
-	bin = ((unsigned long)ref >> 2) % REF_ALLOC_BINS;
+	bin = ((uintptr_t)ref >> 2) % REF_ALLOC_BINS;
 
 	ref->next = ref_list[bin];
 	ref_list[bin] = ref;
@@ -241,7 +241,7 @@ __ref_alloc(size_t len, const char *file, const char *func, int line)
 		ref->file = file;
 		ref->func = func;
 		ref->line = line;
-		guard = (guard_t*)((unsigned long)block +
+		guard = (guard_t*)((uintptr_t)block +
 				   sizeof(refcounter_t) + len);
 		guard->magic = GUARD_MAGIC;
 		ref_add(ref);
@@ -406,7 +406,7 @@ ref_hold(void *p)
 	if (p) {
 #ifdef DEBUG
 		assert(ref->magic == ALLOC_MAGIC);
-		guard = (guard_t*)((unsigned long)block +
+		guard = (guard_t*)((uintptr_t)block +
 				   sizeof(refcounter_t) + ref->length);
 		assert(guard->magic == GUARD_MAGIC);
 #endif /* DEBUG */
@@ -449,7 +449,7 @@ ref_release(void *p)
 			   p, ref, ref->refcount, ref->length);
 #ifdef DEBUG
 		assert(ref->magic == ALLOC_MAGIC);
-		guard = (guard_t*)((unsigned long)block +
+		guard = (guard_t*)((uintptr_t)block +
 				   sizeof(refcounter_t) + ref->length);
 		assert(guard->magic == GUARD_MAGIC);
 #endif /* DEBUG */
